checkpoint: most EC tests pass
